mm-gup_test-start-stop-read-functionality-for-pin-longterm-test-fix
silence a warning on 32-bit

https://lore.kernel.org/all/202210182101.XOyPZEQV-lkp@intel.com/T/#u


Link: https://lkml.kernel.org/r/74adbb51-6e33-f636-8a9c-2ad87bd9007e@redhat.com

+++ mm/gup_test.c
@@ -299,7 +299,8 @@ static inline int pin_longterm_test_read(unsigned long arg)
 	for (i = 0; i < pin_longterm_test_nr_pages; i++) {
 		void *addr = page_to_virt(pin_longterm_test_pages[i]);
 
-		if (copy_to_user((void __user *)user_addr, addr, PAGE_SIZE))
+		if (copy_to_user((void __user *)(unsigned long)user_addr, addr,
+				 PAGE_SIZE))
 			return -EFAULT;
 		user_addr += PAGE_SIZE;
 	}
